@@ -47,13 +47,12 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
-        <junit.jupiter.version>5.0.1</junit.jupiter.version>
-        <junit.platform.version>1.0.1</junit.platform.version>
+        <maven.compiler.source>11</maven.compiler.source>
+        <maven.compiler.target>11</maven.compiler.target>
+        <junit.jupiter.version>5.9.0</junit.jupiter.version>
+        <junit.platform.version>1.9.0</junit.platform.version>
         <junit.vintage.version>4.12.1</junit.vintage.version>
-        <junit.version>4.13.1</junit.version>
-        <surefire.version>2.19.1</surefire.version>
+        <surefire.version>3.0.0-M7</surefire.version>
         <surefireArgLine />
     </properties>
 
