@@ -1,6 +1,6 @@
 package no.maddin.mockjdbc;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.sql.Connection;
 import java.sql.Driver;
